requirements.txt
anyio==4.4.0
asgiref==3.8.1
certifi==2024.7.4
charset-normalizer==3.3.2
Django==5.0.7
django-cors-headers==4.4.0
django-crontab==0.7.1
django-filter==24.2
djangorestframework==3.15.2
drf-yasg==1.21.7
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
inflection==0.5.1
packaging==24.1
pillow==10.4.0
psycopg2-binary==2.9.9
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-telegram-bot==21.4
pytz==2024.1
PyYAML==6.0.1
requests==2.32.3
setuptools==70.3.0
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.1
uritemplate==4.1.1
urllib3==2.2.2
